[GH-#510] Restructuring of Genesis module
