+++ apps/aecore/lib/aecore/chain/block.ex
@@ -3,15 +3,9 @@ defmodule Aecore.Chain.Block do
   Structure of the block
   """
 
-  alias Aecore.Account.Account
-  alias Aecore.Account.AccountStateTree
   alias Aecore.Chain.Block
   alias Aecore.Chain.Header
-  alias Aecore.Chain.Chainstate
   alias Aecore.Tx.SignedTx
-  alias Aecore.Chain.BlockValidation
-  alias Aecore.Governance.GovernanceConstants, as: Governance
-  alias Aeutil.Genesis
   alias Aeutil.Serialization
 
   @version 14
@@ -28,54 +22,6 @@ defmodule Aecore.Chain.Block do
   @spec current_block_version() :: non_neg_integer()
   def current_block_version do
     @version
-  end
-
-  @spec genesis_header() :: Header.t()
-  defp genesis_header do
-    header = %{
-      height: Governance.genesis_height(),
-      prev_hash: Governance.genesis_prev_hash(),
-      txs_hash: Governance.genesis_txs_hash(),
-      root_hash: Chainstate.calculate_root_hash(genesis_populated_trees()),
-      time: Governance.genesis_time(),
-      nonce: Governance.genesis_nonce(),
-      miner: Governance.genesis_miner(),
-      pow_evidence: Governance.genesis_evidence(),
-      version: Governance.genesis_version(),
-      target: Governance.genesis_target()
-    }
-
-    struct(Header, header)
-  end
-
-  def genesis_hash do
-    BlockValidation.block_header_hash(genesis_header())
-  end
-
-  @spec genesis_block() :: Block.t()
-  def genesis_block do
-    header = genesis_header()
-    %Block{header: header, txs: []}
-  end
-
-  def genesis_populated_trees do
-    genesis_populated_trees(Genesis.preset_accounts())
-  end
-
-  def genesis_populated_trees(accounts) do
-    chainstate_init = Chainstate.create_chainstate_trees()
-    miner = {Governance.genesis_miner(), Governance.coinbase_transaction_amount()}
-
-    Enum.reduce([miner | accounts], chainstate_init, fn {pubkey, balance}, new_trees ->
-      new_acounts =
-        new_trees.accounts
-        |> AccountStateTree.put(
-          pubkey,
-          Account.new(%{balance: balance, nonce: 0, pubkey: pubkey})
-        )
-
-      struct(new_trees, accounts: new_acounts)
-    end)
   end
 
   @spec encode_to_map(Block.t()) :: map()
