+++ apps/aehttpserver/config/test.exs
@@ -1,7 +1,7 @@
 use Mix.Config
 
 config :aehttpserver, Aehttpserver.Web.Endpoint,
-  http: [port: 4000],
+  http: [port: 4004],
   server: true
 
 # Print only warnings and errors during test
